refactored plotting module
