+++ docs/modules/plotting.rst
@@ -1,0 +1,10 @@
+.. highlight:: python
+
+==================================
+The plotting (:code:`pl`) accessor
+==================================
+
+The plotting accessor provides functions for visualizing segmentation results, such as cell type proportions and UMAP embeddings.
+
+.. automodule:: segtraq.pl.plotting
+   :members:
