+++ src/segtraq/pl/__init__.py
@@ -1,17 +1,13 @@
-from .plotting_functions import (
-    plot_box_by_celltype,
-    plot_box_by_celltype_combined,
-    plot_celltype_proportions_stacked,
-    plot_mecr_boxplot,
-    plot_umaps_by_feature,
-    save_mask_to_tiff,
+from .plotting import (
+    boxplot,
+    boxplot_combined,
+    celltype_proportions,
+    umap,
 )
 
 __all__ = [
-    "plot_celltype_proportions_stacked",
-    "plot_umaps_by_feature",
-    "plot_box_by_celltype",
-    "plot_box_by_celltype_combined",
-    "plot_mecr_boxplot",
-    "save_mask_to_tiff",
+    "celltype_proportions",
+    "umap",
+    "boxplot",
+    "boxplot_combined",
 ]
